+++ Source/TrinityFlow/Public/Core/ShardComponent.h
@@ -7,6 +7,7 @@
 
 DECLARE_DYNAMIC_MULTICAST_DELEGATE_TwoParams(FOnShardCollected, EShardType, ShardType, int32, NewInactiveCount);
 DECLARE_DYNAMIC_MULTICAST_DELEGATE_ThreeParams(FOnShardsActivated, EShardType, ShardType, int32, ActivatedCount, int32, NewActiveCount);
+DECLARE_DYNAMIC_MULTICAST_DELEGATE_FourParams(FOnDamageBonusChanged, float, SoulBonus, float, PhysicalBonus, float, SoulStanceBonus, float, PhysicalStanceBonus);
 
 USTRUCT(BlueprintType)
 struct FShardData
@@ -52,6 +53,16 @@ public:
 
     UFUNCTION(BlueprintCallable, Category = "Shards")
     void GetShardCounts(int32& OutSoulActive, int32& OutSoulInactive, int32& OutPowerActive, int32& OutPowerInactive) const;
+    
+    // Damage Bonus Calculations
+    UFUNCTION(BlueprintCallable, Category = "Shards|Damage")
+    float GetSoulDamageBonus() const;
+    
+    UFUNCTION(BlueprintCallable, Category = "Shards|Damage")
+    float GetPhysicalDamageBonus() const;
+    
+    UFUNCTION(BlueprintCallable, Category = "Shards|Damage")
+    void GetDamageBonuses(float& OutSoulBonus, float& OutPhysicalBonus, float& OutSoulStanceBonus, float& OutPhysicalStanceBonus) const;
 
     // Events
     UPROPERTY(BlueprintAssignable, Category = "Shards")
@@ -59,6 +70,9 @@ public:
 
     UPROPERTY(BlueprintAssignable, Category = "Shards")
     FOnShardsActivated OnShardsActivated;
+    
+    UPROPERTY(BlueprintAssignable, Category = "Shards")
+    FOnDamageBonusChanged OnDamageBonusChanged;
 
 protected:
     UPROPERTY(VisibleAnywhere, BlueprintReadOnly, Category = "Shards")
@@ -67,6 +81,10 @@ protected:
     // Maximum shards that can be activated at once at an altar
     UPROPERTY(EditDefaultsOnly, BlueprintReadOnly, Category = "Shards")
     int32 MaxShardsPerActivation = 5;
+    
+    // Damage bonus per shard (3%)
+    UPROPERTY(EditDefaultsOnly, BlueprintReadOnly, Category = "Shards")
+    float DamageBonusPerShard = 0.03f;
 
 private:
     void InitializeShardData();